refactor: 썸내일 그리드 공통 컴포넌트 폴더로 이동
