+++ src/components/common/Thumbnail.stories.tsx
