+++ src/components/common/Thumbnail.tsx
