+++ src/components/common/ThumbnailGrid.stories.tsx
